Примитивы теперь являются сущностями
