+++ src/main/java/amaralus/apps/hackandslash/graphics/entities/primitives/Primitive.java
@@ -1,6 +1,7 @@
 package amaralus.apps.hackandslash.graphics.entities.primitives;
 
 import amaralus.apps.hackandslash.graphics.entities.Color;
+import amaralus.apps.hackandslash.graphics.entities.RenderComponent;
 import amaralus.apps.hackandslash.graphics.entities.gpu.FloatVertexBufferObject;
 import amaralus.apps.hackandslash.graphics.entities.gpu.VertexArraysObject;
 
@@ -8,7 +9,7 @@ import java.nio.FloatBuffer;
 
 import static amaralus.apps.hackandslash.utils.VectMatrUtil.toArray;
 
-public abstract class Primitive {
+public abstract class Primitive implements RenderComponent {
 
     private final VertexArraysObject vao;
     protected Color color;
@@ -22,6 +23,11 @@ public abstract class Primitive {
         this.color = color;
         var vbo = getVbo(1);
         vbo.updateBuffer(FloatBuffer.wrap(toArray(color.getVector())), 0);
+    }
+
+    @Override
+    public void update(long elapsedTime) {
+        // Примитивам пока не нужно обновляться
     }
 
     protected FloatVertexBufferObject getVbo(int index) {
